Add api folder and change package.json
